refactor(MAJOR): Move lib into another crate. Use cargo workspaces.
# What?

Enclose the application logic in a separate crate.

# Why?

To run experiments it is convenient to build a separate bin to perform a
lsp-command to all classes which satisfy a constraint and exit after
execution.

This change allows to build separately the language server and any short
lived binaries for experiments and profiling.